+++ src/LaravelConcurrentLimiterServiceProvider.php
@@ -1,7 +1,10 @@
 <?php
 
-namespace Patrocle\LaravelConcurrentLimiter;
+declare(strict_types=1);
 
+namespace Largerio\LaravelConcurrentLimiter;
+
+use Largerio\LaravelConcurrentLimiter\Contracts\ConcurrentLimiter;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
 
@@ -14,9 +17,16 @@ class LaravelConcurrentLimiterServiceProvider extends PackageServiceProvider
             ->hasConfigFile();
     }
 
+    public function packageRegistered(): void
+    {
+        $this->app->singleton(ConcurrentLimiter::class, LaravelConcurrentLimiter::class);
+        $this->app->singleton(LaravelConcurrentLimiter::class);
+    }
+
     public function packageBooted(): void
     {
-        $router = $this->app['router'];
+        /** @var \Illuminate\Routing\Router $router */
+        $router = $this->app->make('router');
         $router->aliasMiddleware('concurrent.limit', LaravelConcurrentLimiter::class);
     }
 }