+++ tests/Pest.php
@@ -1,5 +1,5 @@
 <?php
 
-use Patrocle\LaravelConcurrentLimiter\Tests\TestCase;
+use Largerio\LaravelConcurrentLimiter\Tests\TestCase;
 
 uses(TestCase::class)->in(__DIR__);